+++ test/parser/unparse.spec.ts
@@ -597,4 +597,18 @@ describe('whitespaces', () => {
 
     expect(unparsed).toEqual(formula)
   })
+
+  it('when ignoreWhiteSpace = \'any\', should unparse a non-breakable space character', () => {
+    const config = new Config({ ignoreWhiteSpace: 'any' })
+    const lexerConfig = buildLexerConfig(config)
+    const parser = buildEmptyParserWithCaching(config, sheetMapping)
+    const unparser = new Unparser(config, lexerConfig, sheetMapping.fetchDisplayName, new NamedExpressions())
+
+    const formula = '=\u00A01'
+    const ast = parser.parse(formula, adr('A1')).ast
+
+    const unparsed = unparser.unparse(ast, adr('A1'))
+
+    expect(unparsed).toEqual(formula)
+  })
 })